moved modal to correct dir
